Move GridData struct and related to new file
