+++ src/griddata.cpp
@@ -1,0 +1,51 @@
+#include "griddata.h"
+#include "controller.h"
+#include "misc.h"
+
+// GridCol definitions
+int GridCol::getNumberRows(const bool include_header) const {
+  return values.size() + (include_header? 2 : 0);
+}
+
+someText GridCol::getElem(const int row, const bool include_header) const {
+  if (include_header) {
+    if (row == 0){
+      return someText(header);
+    }
+    else if (row == 1){
+      return someText(unit);
+    }
+  }
+  return someText(values[row - (include_header? 2 : 0)]);
+}
+
+void GridCol::pushBack(const std::string val){
+  values.push_back(val);
+}
+
+// GridData definitions
+size_t GridData::getNumberRows(bool include_header) const {return columns[0].getNumberRows(include_header);}
+
+void GridData::print() const {
+  constexpr int width = 20;
+  
+  for (size_t row = 0; row < getNumberRows(true); ++row){
+    for (const GridCol& col : columns) {
+      if (col.hide_col){continue;}
+      someText cell = col.getElem(row,true);
+      if (cell.str.empty()) {
+        Ctrl::getInstance()->notifyUser(wfield(width, cell.wstr));
+      }
+      else{
+        Ctrl::getInstance()->notifyUser(field(width, cell.str));
+      }
+    }
+    Ctrl::getInstance()->notifyUser("\n");
+  }
+}
+
+void GridData::storeValues(const std::vector<std::string> vals){
+  for (size_t col = 0; col < vals.size(); ++col){
+    columns[col].pushBack(vals[col]);
+  }
+}
